+++ frontend/src/app/app.tsx
@@ -1,4 +1,4 @@
-import React, {useRef, useState, useCallback} from 'react';
+import React, {useRef, useState, useCallback, useMemo} from 'react';
 import './app.scss';
 import SourceSelector from "../component/source-selector/source-selector";
 import PlaylistViewer, {IPlaylistViewer} from "../component/playlist-viewer/playlist-viewer";
@@ -8,6 +8,8 @@ import {PlaylistGroup} from "../model/playlist";
 import {useServices} from "../provider/service-provider";
 import {first} from "rxjs/operators";
 import Progress from '../component/progress/progress';
+import PlaylistFilter from "../component/playlist-filter/playlist-filter";
+import {Subject} from "rxjs";
 
 /* eslint-disable @typescript-eslint/no-empty-interface */
 interface AppProps {
@@ -15,17 +17,18 @@ interface AppProps {
 }
 
 export default function App(props: AppProps) {
+    const searchChannel = useMemo<Subject<string>>(() => new Subject<string>(), []);
     const [progress, setProgress] = useState<boolean>(false);
-    const [playlist, setPlaylist] = useState<PlaylistGroup[]>(null);
+    const [playlist, setPlaylist] = useState<PlaylistGroup[]>([]);
     const viewerRef = useRef<IPlaylistViewer>();
     const {enqueueSnackbar/*, closeSnackbar*/} = useSnackbar();
     const services = useServices();
     const handleDownload = useCallback((url: string) => {
         setProgress(true);
         services.playlist().getPlaylist(url).pipe(first()).subscribe({
-            next: (playlist: PlaylistGroup[]) => {
+            next: (pl: PlaylistGroup[]) => {
                 enqueueSnackbar('Sucessfully downloaded playlist', {variant: 'success'})
-                setPlaylist(playlist);
+                setPlaylist(pl);
             },
             error: (err) => {
                 setProgress(false);
@@ -42,11 +45,21 @@ export default function App(props: AppProps) {
         }
     }, [services]);
 
+
+    const handleFilter = useCallback((filter: string): void => {
+        searchChannel.next(filter);
+    }, [searchChannel]);
+
+    const handleProgress = useCallback((value: boolean) => {
+       setProgress(value);
+    }, []);
+
     return (
         <div className="app">
             <div className={'app-header'}>m3u-filter</div>
             <SourceSelector onDownload={handleDownload}/>
-            <PlaylistViewer ref={viewerRef} playlist={playlist}/>
+            <PlaylistFilter onFilter={handleFilter}/>
+            <PlaylistViewer ref={viewerRef} playlist={playlist} searchChannel={searchChannel} onProgress={handleProgress}/>
             <Toolbar onDownload={handleSave}/>
             <Progress visible={progress}/>
         </div>