+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## v0.7.0(2022-01-20)
+* Updated frontend libraries
+* Added Search input
+
 ## v0.6.0(2021-12-29)
 * Added options to target, currently only ignore_logo
 * Added sorting to groups